updated to wc with fixed RSE

+++ Version5_wc/Ganlib/src/filmod.f90
@@ -57,7 +57,7 @@
 module FILMOD
    use, intrinsic :: iso_c_binding
    type, bind(c) :: FIL_file
-      integer :: unit
+      integer(c_int) :: unit
       type(c_ptr) :: kdi_file
    end type FIL_file
 contains
